+++ dev_db_builder.ps1
@@ -1,0 +1,6 @@
+Import-Module ./Update-TRTDatabase.ps1 -Force
+$server = '0.0.0.0'
+$database = 'trtdb'
+$userName = 'root'
+$loginPath = 'trtdb'
+Update-TRTDatabase -loginPath $loginPath -serverName $server -databaseName $database -userName $userName -updateCredentials