+++ stored_procs/getUserFeaturesByGUID.sql
@@ -2,7 +2,7 @@ DELIMITER $$
 
 DROP procedure IF EXISTS `getUserFeaturesByGUID`;
 
-CREATE DEFINER=`root`@`localhost` PROCEDURE `getUserFeaturesByGUID`(
+CREATE PROCEDURE `getUserFeaturesByGUID`(
 	IN target_guid VARCHAR(255)
 ) 
 BEGIN