+++ stored_procs/increaseUserPopulationByGUID.sql
@@ -2,7 +2,7 @@ DELIMITER $$
 
 DROP procedure IF EXISTS `increaseUserPopulationByGUID`;
 
-CREATE DEFINER=`root`@`localhost` PROCEDURE `increaseUserPopulationByGUID`(
+CREATE PROCEDURE `increaseUserPopulationByGUID`(
 	IN target_guid VARCHAR(255)
 ) 
 BEGIN