+++ stored_procs/setUserPopulationByGUID.sql
@@ -2,7 +2,7 @@ DELIMITER $$
 
 DROP procedure IF EXISTS `setUserPopulationByGUID`;
 
-CREATE DEFINER=`root`@`localhost` PROCEDURE `setUserPopulationByGUID`(
+CREATE PROCEDURE `setUserPopulationByGUID`(
 	IN target_guid VARCHAR(255),
 	IN new_population INT(11)
 )