+++ tables/003_feature_type.sql
@@ -1,4 +1,4 @@
-CREATE TABLE `feature_type` (
+CREATE TABLE IF NOT EXISTS `feature_type` (
   `feature_id` int(11) NOT NULL AUTO_INCREMENT,
   `feature_name` varchar(255) NOT NULL,
   PRIMARY KEY (`feature_id`),